Added confirmation before the user selects to delete the accounts that they have choosen

+++ components/data-table.tsx
@@ -22,6 +22,7 @@ import {
     TableRow,
 } from "@/components/ui/table"
 
+import { useConfirm } from "@/hooks/use-confirm"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
 import { Trash } from "lucide-react"
@@ -44,6 +45,10 @@ export function DataTable<TData, TValue>({
     disabled,
 
 }: DataTableProps<TData, TValue>) {
+    const [ConfirmDialog, confirm] = useConfirm (
+        "Are you sure?",
+        "You are about to preform a bulk delete."
+    )
     const [sorting, setSorting] = React.useState<SortingState>([])
     const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
         []
@@ -70,6 +75,7 @@ export function DataTable<TData, TValue>({
 
 return (
     <div>
+        <ConfirmDialog/>
         <div className="flex items-center py-4">
             <Input
             placeholder={`Filter ${filterKey}...`}
@@ -84,9 +90,14 @@ return (
                   size="sm"
                   variant="outline"
                   className="ml-auto font-normal text-xs"
-                  onClick={() => {
-                    onDelete(table.getFilteredSelectedRowModel().rows)
+                  onClick={ async () => {
+                    const ok = await confirm();
+
+                    if(ok){
+                        onDelete(table.getFilteredSelectedRowModel().rows)
                     table.resetRowSelection()
+
+                    }
                   }}
                 >
                     <Trash className="size-4 mr-2"/>
